+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Upgrade `sea-query-binder` to `0.4` https://github.com/SeaQL/sea-schema/pull/104
 * Replace the use of `SeaRc<T>` where `T` isn't `dyn Iden` with `RcOrArc<T>` https://github.com/SeaQL/sea-schema/pull/107
 
+### Breaking changes
+
+* API now returns `Result` instead of panic on errors https://github.com/SeaQL/sea-schema/pull/109
+
 ## 0.11.0 - 2023-01-05
 
 * Upgrade SeaQuery to 0.28 https://github.com/SeaQL/sea-schema/pull/90